# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Toyo Kambocha HRMS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /employees:
    get:
      operationId: listEmployees
      tags: [employees]
      summary: List all employees
      parameters:
        - name: department
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of employees
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployeeListResponse"
    post:
      operationId: createEmployee
      tags: [employees]
      summary: Create employee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEmployeeRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
  /employees/{id}:
    get:
      operationId: getEmployee
      tags: [employees]
      summary: Get employee by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Employee
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
    put:
      operationId: updateEmployee
      tags: [employees]
      summary: Update employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEmployeeRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
    delete:
      operationId: deleteEmployee
      tags: [employees]
      summary: Delete employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /leaves:
    get:
      operationId: listLeaves
      tags: [attendance]
      summary: List leave requests
      parameters:
        - name: employeeId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Leave requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeaveRequest"
    post:
      operationId: applyLeave
      tags: [attendance]
      summary: Apply for leave
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplyLeaveRequest"
      responses:
        "201":
          description: Leave applied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaveRequest"
  /leaves/{id}/status:
    put:
      operationId: updateLeaveStatus
      tags: [attendance]
      summary: Approve or reject leave
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLeaveStatusRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaveRequest"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
...
    Employee:
      type: object
      properties:
        id:
          type: integer
        employeeCode:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
          nullable: true
        avatar:
          type: string
          nullable: true
        departmentId:
          type: integer
          nullable: true
        departmentName:
          type: string
          nullable: true
        designation:
          type: string
          nullable: true
        role:
          type: string
          enum: [admin, hr, manager, employee]
        status:
          type: string
          enum: [active, inactive, on_leave]
        joinDate:
          type: string
          nullable: true
        salary:
          type: number
          nullable: true
        managerId:
          type: integer
          nullable: true
        managerName:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        dateOfBirth:
          type: string
          nullable: true
        gender:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, employeeCode, firstName, lastName, email, role, status, createdAt]
    EmployeeListResponse:
      type: object
      properties:
        employees:
          type: array
          items:
            $ref: "#/components/schemas/Employee"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [employees, total, page, limit]
    CreateEmployeeRequest:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
          nullable: true
        departmentId:
          type: integer
          nullable: true
        designation:
          type: string
          nullable: true
        role:
          type: string
          enum: [admin, hr, manager, employee]
        joinDate:
          type: string
          nullable: true
        salary:
          type: number
          nullable: true
        managerId:
          type: integer
          nullable: true
        address:
          type: string
          nullable: true
        dateOfBirth:
          type: string
          nullable: true
        gender:
          type: string
          nullable: true
        password:
          type: string
      required: [firstName, lastName, email, role, password]
    UpdateEmployeeRequest:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
          nullable: true
        departmentId:
          type: integer
          nullable: true
        designation:
          type: string
          nullable: true
        role:
          type: string
          enum: [admin, hr, manager, employee]
        status:
          type: string
          enum: [active, inactive, on_leave]
        joinDate:
          type: string
          nullable: true
        salary:
          type: number
          nullable: true
        managerId:
          type: integer
          nullable: true
        address:
          type: string
          nullable: true
        dateOfBirth:
          type: string
          nullable: true
        gender:
          type: string
          nullable: true
...
    LeaveRequest:
      type: object
      properties:
        id:
          type: integer
        employeeId:
          type: integer
        employeeName:
          type: string
        leaveType:
          type: string
          enum: [annual, sick, casual, maternity, paternity, unpaid]
        startDate:
          type: string
        endDate:
          type: string
        days:
          type: number
        reason:
          type: string
        status:
          type: string
          enum: [pending, approved, rejected]
        approvedBy:
          type: integer
          nullable: true
        approvedByName:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, employeeId, employeeName, leaveType, startDate, endDate, days, reason, status, createdAt]
    ApplyLeaveRequest:
      type: object
      properties:
        employeeId:
          type: integer
        leaveType:
          type: string
          enum: [annual, sick, casual, maternity, paternity, unpaid]
        startDate:
          type: string
        endDate:
          type: string
        reason:
          type: string
      required: [employeeId, leaveType, startDate, endDate, reason]
    UpdateLeaveStatusRequest:
      type: object
      properties:
        status:
          type: string
          enum: [approved, rejected]
        approvedBy:
          type: integer
      required: [status, approvedBy]
...